+++ cs3560cli/commands/watch_zip.py
@@ -36,6 +36,7 @@ def is_7z_available() -> bool:
             return False
         except subprocess.CalledProcessError:
             return False
+    return False
 
 
 def extract(path: Path):