+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
+- Fix not being able to `Ctrl-C` the `blackboard student-list` command on Windows.
+- Fix mypy's errors/warnings.
+
 ## v0.2.1a1
 
 - Add `blackboard categorize` command from `krerkkiat/ta-tooling`.
 - Add `github get-team-id` command from `OU-CS3560/org-bulk-invite`.
-- Fix not being able to `Ctrl-C` the `blackboard student-list` command on Windows.
 
 ## v0.2.0